third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/object-embed-tag-expected.txt

CONSOLE ERROR: line 8: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?wait-for-load=1&q=%3Cobject%20id=%27object%27%20name=%27plugin%27%20type=%27application/x-blink-test-plugin%27%3E%3Cparam%20name=%27movie%27%20value=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3Cembed%20id=%27embed%27%20name=%27plugin%27%20type=%27application/x-blink-test-plugin%27%20src=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3C/object%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
CONSOLE ERROR: line 8: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?wait-for-load=1&q=%3Cobject%20id=%27object%27%20name=%27plugin%27%20type=%27application/x-blink-test-plugin%27%3E%3Cparam%20name=%27movie%27%20value=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3Cembed%20id=%27embed%27%20name=%27plugin%27%20type=%27application/x-blink-test-plugin%27%20src=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3C/object%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
CONSOLE ERROR: line 8: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?wait-for-load=1&q=%3Cobject%20id=%27object%27%20name=%27plugin%27%20type=%27application/x-blink-test-plugin%27%3E%3Cparam%20name=%27movie%27%20value=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3Cembed%20id=%27embed%27%20name=%27plugin%27%20type=%27application/x-blink-test-plugin%27%20src=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%27%20/%3E%3C/object%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
